requirements.txt
matplotlib==3.4.3
numpy==1.21.4
pandas==1.3.4
Pillow==9.3.0
plotly==5.4.0
scikit_learn==1.0.2
streamlit==1.8.1
streamlit_option_menu==0.3.2
